refactor: extract duplicate migration logic to shared utility
Changes:
1. Created src/process/bridge/migrationUtils.ts with shared migrateConversationToDatabase function
2. Updated conversationBridge.ts to use shared migration utility
3. Updated databaseBridge.ts to use shared migration utility

Benefits:
- Eliminates duplicate code (88 lines removed, 56 lines added)
- Single source of truth for migration logic
- Easier to maintain and update migration behavior
- Consistent migration behavior across all bridge modules
